Add PWM based audio DAC
This change adds a PWM based audio DAC as an alternative to the Gameduino's existing delta-sigma converter.

The pwm_generator module can be (and is) set to create multiple shorter pulses per sample, using the WIDTH_BITS parameter.
The sampling time is reduced below 2^BITS samples using the SAMPLE_PERIOD_BITS parameter.
In both cases, the module uses dithering over a period of 2^BITS cycles to make use of the full bit width of the input samples.

Changes:
* Add module pwm_generator
* Add define USE_PWM_DAC to replace the default audio DAC
* When USE_PWM_DAC is defined:
    * Use 16 bit audio data instead of 13 bit
    * Use the pwm_generator module's ready output to trigger FMicrosynth, instead of a dedicated counter